meds: made storage persistent.

+++ services/meds/storage.c
@@ -1,5 +1,13 @@
 #include "storage.h"
 
+enum store_flags {
+	ST_NONE       = 0x00,
+	ST_PROTECT    = 0x01, 
+	ST_PERSIST    = 0x02, 
+	ST_SAFE       = 0x04,
+	ST_ADD_RECENT = 0x08
+};
+
 struct tree_node {
 	int my_node;
 	uuid_t key;
@@ -18,9 +26,11 @@ int items_count;
 uuid_t recent_keys[TREE_MAXITEMS];
 int most_recent_key;
 
-char * _store_item(const uuid_t key, const char *value, bool protect);
+const char* storage_path;
 
-char *render_uuid(const uuid_t uuid) {
+char * store_item_internal(const uuid_t key, const char *value, enum store_flags flags);
+
+char *render_uuid(const uuid_t uuid) { // debug only
 	char *s = malloc(37);
 	uuid_unparse_lower(uuid, s);
 	return s;
@@ -40,8 +50,12 @@ struct tree_node* get_item(int index) {
 	return &data.tree_items[index - 1];
 }
 
+bool is_item_protected(struct tree_node* item) {
+	return item->protected;
+}
+
 bool is_protected(int node) {
-	return get_item(data.tree[node])->protected;
+	return is_item_protected(get_item(data.tree[node]));
 }
 
 const uuid_t max_uuid = { 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff, 0xff };
@@ -73,11 +87,11 @@ void prefill_storage(int depth, const uuid_t lower, const uuid_t upper) {
 	while (front < back) {
 		uuid_t avg;
 		get_average_uuid(queue[front].lower, queue[front].upper, avg);
-		store_item(avg, "X", true);
+		store_item_internal(avg, "X", ST_PROTECT | ST_ADD_RECENT);
 
 		if (queue[front].depth > 0) {
 			if (back + 2 > TREE_MAXITEMS) {
-				perror("tried to insert too many items\n");
+				fprintf(stderr, "Tried to insert too many items\n");
 				exit(1);
 			}
 
@@ -97,17 +111,79 @@ void prefill_storage(int depth, const uuid_t lower, const uuid_t upper) {
 }
 
 
+void save_to_file() {
+	int fd;
+
+	if ((fd = open(storage_path, O_WRONLY | O_CREAT, 0666)) < 0)
+	{
+		perror("Can't open storage\n");
+		exit(1);
+	}
+
+	for (int i = 1; i <= TREE_MAXITEMS; i++) {
+		struct tree_node* item = get_item(i);
+		if (is_item_empty(item) || is_item_protected(item))
+			continue;
+		if (write(fd, item, sizeof(struct tree_node)) < 0)
+		{
+			perror("Failed to write item!");
+			exit(1);
+		}
+	}
+
+	close(fd);
+}
+
+void load_from_file() {
+	int fd;
+
+	if ((fd = open(storage_path, O_RDONLY, 0666)) < 0)
+	{
+		if (errno == ENOENT)
+			return;
+		perror("Can't open storage\n");
+		exit(1);
+	}
+
+	int items_read = 0;
+	struct tree_node item;
+	while (true)
+	{
+		int bytesRead = read(fd, &item, sizeof(struct tree_node));
+
+		if (!bytesRead)
+			break;
+
+		if (bytesRead < sizeof(item))
+		{
+			printf("Item %d is corrupt\n", items_read);
+			break;
+		}
+
+		store_item_internal(item.key, item.value, ST_SAFE | ST_ADD_RECENT);
+		items_read++;
+	}
+
+	printf("Loaded %d items.\n", items_count);
+
+	close(fd);
+}
+
+
 void init_storage(const char* file_path) {
 	bzero(data.tree, sizeof(data.tree));
 	bzero(data.tree_items, sizeof(data.tree_items));
 	next_item = 0;
 	items_count = 0;
 	most_recent_key = 0;
+	storage_path = file_path;
 
 	uuid_t lower;
 	bzero(lower, sizeof(uuid_t));
 	prefill_storage(10, lower, max_uuid);
 	DEBUG("!! prefill inserted %d nodes!\n", items_count);
+
+	load_from_file();
 }
 
 int find_node(const uuid_t key, int root) {
@@ -127,7 +203,7 @@ int find_node(const uuid_t key, int root) {
 }
 
 void remove_children(int root, struct tree_node* children, int* children_count) {
-	if (is_empty(root))
+	if (is_empty(root) || root >= TREE_MAXNODES)
 		return;
 
 	struct tree_node* item = get_item(data.tree[root]);
@@ -161,13 +237,13 @@ void delete_item(int root) {
 	qsort(children, children_count, sizeof(struct tree_node), child_comp);
 
 	int mid = children_count / 2;
-	_store_item(children[mid].key, children[mid].value, false);
+	store_item_internal(children[mid].key, children[mid].value, ST_NONE);
 	for (int i = mid + 1; i < children_count; i++) {
-		_store_item(children[2 * mid - i].key, children[2 * mid - i].value, false);
-		_store_item(children[i].key, children[i].value, false);
+		store_item_internal(children[2 * mid - i].key, children[2 * mid - i].value, ST_NONE);
+		store_item_internal(children[i].key, children[i].value, ST_NONE);
 	}
 	if (children_count % 2 == 0)
-		_store_item(children[0].key, children[0].value, false);
+		store_item_internal(children[0].key, children[0].value, ST_NONE);
 }
 
 int16_t allocate_item() {
@@ -186,41 +262,54 @@ int find_oldest_node() {
 	for (int i = 0; i <= TREE_MAXITEMS; i++) {
 		int idx = (most_recent_key + i) % TREE_MAXITEMS;
 		int node = find_node(recent_keys[idx], 0);
+		// DEBUG("!! checking item = %d, node = %d, key = %s\n", idx, node, render_uuid(recent_keys[idx]));
 		if (is_empty(node) || is_protected(node))
 			continue;
 		DEBUG("!! oldest node is %s, i = %d, mrk = %d\n", render_uuid(recent_keys[idx]), i, most_recent_key);
 		return node;
 	}
-	perror("no node to delete");
+	fprintf(stderr, "No node to delete\n");
 	exit(1);
 }
 
-char * _store_item(const uuid_t key, const char *value, bool protect) {
+char * store_item_internal(const uuid_t key, const char *value, enum store_flags flags) {
+	if (strlen(value) == 0)
+		return 0;
+
 	int node = find_node(key, 0);
+	if (flags & ST_SAFE && node >= TREE_MAXNODES)
+		return 0;
 
 	if (is_empty(node)) {
 		if (items_count == TREE_MAXITEMS) {
 			delete_item(find_oldest_node());
-			return _store_item(key, value, protect);
+			return store_item_internal(key, value, flags);
 		}
 		data.tree[node] = allocate_item();
 	}
 
-	DEBUG("!! storing %s to item %d (node %d)\n", render_uuid(key), data.tree[node], node);
+	DEBUG("!! storing %s to item %d (node %d), flags = %x\n", render_uuid(key), data.tree[node], node, flags);
 
 	struct tree_node* item = get_item(data.tree[node]);
 	item->my_node = node;
-	item->protected |= protect;
+	item->protected |= (flags & ST_PROTECT) > 0;
 	memcpy(item->key, key, sizeof(uuid_t));
-	return strcpy(item->value, value);
+	char* result = strcpy(item->value, value);
+
+	if (flags & ST_ADD_RECENT) {
+		DEBUG("!! outer storing %s @ mrk = %d\n", render_uuid(key), most_recent_key);
+		memcpy(recent_keys[most_recent_key], key, sizeof(uuid_t));
+		most_recent_key = (most_recent_key + 1) % TREE_MAXITEMS;
+	}
+
+	if (flags & ST_PERSIST)
+		save_to_file();
+
+	return result;
 }
 
-char * store_item(const uuid_t key, const char *value, bool protect) {
-	_store_item(key, value, protect);
-
-	DEBUG("!! outer storing %s @ mrk = %d\n", render_uuid(key), most_recent_key);
-	memcpy(recent_keys[most_recent_key], key, sizeof(uuid_t));
-	most_recent_key = (most_recent_key + 1) % TREE_MAXITEMS;
+char * store_item(const uuid_t key, const char *value) {
+	return store_item_internal(key, value, ST_PERSIST | ST_ADD_RECENT);
 }
 
 char * load_item(const uuid_t key, char *buffer) {
@@ -251,16 +340,4 @@ int get_tree_height(int id) {
 	if (id >= TREE_MAXNODES || is_empty(id))
 		return 0;
 	return 1 + max(get_tree_height(2 * id + 1), get_tree_height(2 * id + 2)); 
-}
-
-void dump_mem() {
-	printf("tree start: %p\n", &data.tree);
-	printf("tree end: %p\n", &data.tree[TREE_MAXNODES]);
-	printf("tree_items start: %p\n", data.tree_items);
-
-	char* start = (char *)&data.tree[TREE_MAXNODES + 12];
-	for (int i = 0; i < 1024; i++) {
-		printf("%02x ", (uint32_t)start[i]);
-	}
-	printf("\n");
 }
